Merges module tests by component
