--- src/Tablist/modules/automaticActivation.test.js
@@ -1,157 +1,0 @@
-/* eslint-disable max-len */
-import Tablist from '..';
-import { AutomaticActivation } from '.';
-import { events } from '../../../.jest/events';
-
-const {
-  keydownTab,
-  keydownArrowLeft,
-  keydownArrowRight,
-  keydownArrowDown,
-  keydownHome,
-  keydownEnd,
-} = events;
-
-// Set up our document body
-document.body.innerHTML = `
-  <ul class="tablist">
-    <li><a aria-controls="first-panel" href="#first-panel"></a></li>
-    <li><a aria-controls="second-panel" href="#second-panel"></a></li>
-    <li><a aria-controls="third-panel" href="#third-panel"></a></li>
-  </ul>
-  <div id="first-panel" class="panel">
-    <h1>The Article Title</h1>
-    <p>Lorem ipsum dolor sit amet, <a href="first-panel.com/first">consectetur</a>
-  adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna
-  aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris
-  nisi ut aliquip ex ea <a href="first-panel.com/second">commodo consequat</a>. Duis aute irure dolor in
-  reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-  pariatur. Excepteur <a href="first-panel.com/third">sint occaecat</a> cupidatat non proident, sunt in culpa qui
-  officia deserunt mollit anim id est laborum.</p>
-  </div>
-  <div id="second-panel" class="panel">
-    <h1>The Article Title</h1>
-    <p>Lorem ipsum dolor sit amet, <a href="example.com/second">consectetur</a>
-    adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna
-    aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris
-    nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in
-    reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-    pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui
-    officia deserunt mollit anim id est laborum.</p>
-  </div>
-  <div id="third-panel" class="panel">
-    <h1>The Article Title</h1>
-    <p>Lorem ipsum dolor sit amet, <a href="example.com/third">consectetur</a>
-    adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna
-    aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris
-    nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in
-    reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-    pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui
-    officia deserunt mollit anim id est laborum.</p>
-  </div>
-`;
-
-const tabs = document.querySelector('.tablist');
-const panels = document.querySelectorAll('.panel');
-
-// Cached selectors.
-const firstTab = document.querySelector('[aria-controls="first-panel"]');
-const secondTab = document.querySelector('[aria-controls="second-panel"]');
-const thirdTab = document.querySelector('[aria-controls="third-panel"]');
-
-const firstPanel = document.querySelector('#first-panel');
-const secondPanel = document.querySelector('#second-panel');
-const thirdPanel = document.querySelector('#third-panel');
-
-const tablist = new Tablist(tabs, { modules: [AutomaticActivation] });
-
-test('Panels initialize with expected tabindex', () => {
-  expect(firstPanel.getAttribute('tabindex')).toEqual('0');
-  expect(secondPanel.getAttribute('tabindex')).toBeNull();
-  expect(thirdPanel.getAttribute('tabindex')).toBeNull();
-});
-
-test('Arrow keys activate tabs', () => {
-  firstTab.focus();
-  firstTab.dispatchEvent(keydownArrowRight);
-
-  expect(firstTab.getAttribute('aria-selected')).toBeNull();
-  expect(secondTab.getAttribute('aria-selected')).toEqual('true');
-  expect(thirdTab.getAttribute('aria-selected')).toBeNull();
-
-  secondTab.dispatchEvent(keydownArrowRight);
-
-  expect(firstTab.getAttribute('aria-selected')).toBeNull();
-  expect(secondTab.getAttribute('aria-selected')).toBeNull();
-  expect(thirdTab.getAttribute('aria-selected')).toEqual('true');
-
-  thirdTab.dispatchEvent(keydownArrowRight);
-
-  expect(firstTab.getAttribute('aria-selected')).toEqual('true');
-  expect(secondTab.getAttribute('aria-selected')).toBeNull();
-  expect(thirdTab.getAttribute('aria-selected')).toBeNull();
-
-  thirdTab.dispatchEvent(keydownArrowLeft);
-
-  expect(firstTab.getAttribute('aria-selected')).toBeNull();
-  expect(secondTab.getAttribute('aria-selected')).toEqual('true');
-  expect(thirdTab.getAttribute('aria-selected')).toBeNull();
-
-  secondTab.dispatchEvent(keydownArrowLeft);
-
-  expect(firstTab.getAttribute('aria-selected')).toEqual('true');
-  expect(secondTab.getAttribute('aria-selected')).toBeNull();
-  expect(thirdTab.getAttribute('aria-selected')).toBeNull();
-
-  firstTab.dispatchEvent(keydownArrowLeft);
-
-  expect(firstTab.getAttribute('aria-selected')).toBeNull();
-  expect(secondTab.getAttribute('aria-selected')).toBeNull();
-  expect(thirdTab.getAttribute('aria-selected')).toEqual('true');
-});
-
-test('The Home key moves activates the first tab', () => {
-  tablist.switchTo(1);
-
-  secondTab.focus();
-  secondTab.dispatchEvent(keydownHome);
-
-  expect(firstTab.getAttribute('aria-selected')).toEqual('true');
-  expect(secondTab.getAttribute('aria-selected')).toBeNull();
-  expect(thirdTab.getAttribute('aria-selected')).toBeNull();
-});
-
-test('The End key moves activates the last tab', () => {
-  tablist.switchTo(1);
-
-  secondTab.focus();
-  secondTab.dispatchEvent(keydownEnd);
-
-  expect(firstTab.getAttribute('aria-selected')).toBeNull();
-  expect(secondTab.getAttribute('aria-selected')).toBeNull();
-  expect(thirdTab.getAttribute('aria-selected')).toEqual('true');
-});
-
-test('The Tab key moves focus to the active tab panel', () => {
-  tablist.switchTo(2);
-  thirdTab.dispatchEvent(keydownTab);
-  expect(document.activeElement).toEqual(thirdPanel);
-});
-
-test('The active tab panel is added to the tab index', () => {
-  tablist.switchTo(0);
-  expect(firstPanel.getAttribute('tabindex')).toEqual('0');
-  expect(secondPanel.getAttribute('tabindex')).toBeNull();
-  expect(thirdPanel.getAttribute('tabindex')).toBeNull();
-
-  tablist.switchTo(2);
-  expect(firstPanel.getAttribute('tabindex')).toBeNull();
-  expect(secondPanel.getAttribute('tabindex')).toBeNull();
-  expect(thirdPanel.getAttribute('tabindex')).toEqual('0');
-});
-
-test('All DOM attributes are removed from elements managed by this component', () => {
-  tablist.destroy();
-
-  Array.from(panels).forEach((panel) => expect(panel.getAttribute('tabindex')).toBeNull());
-});
